2024-03-09 Build and add Device Reference Synchronization
Build and add Device Reference Synchronization

+++ SensorizMonitoring/Models/PosModel.cs
@@ -4,8 +4,8 @@
     {
         public int type { get; set; }
         public long time { get; set; }
-        public float lat { get; set; }
-        public float lon { get; set; }
-        public float cep { get; set; } }
+        public double lat { get; set; }
+        public double lon { get; set; }
+        public double cep { get; set; } }
 }
