+++ SensorizMonitoring/Models/VibrationModel.cs
@@ -2,8 +2,8 @@
 {
     public class VibrationModel
     {
-        public float x { get; set; }
-        public float y { get; set; }
-        public float z { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
+        public double z { get; set; }
     }
 }
